requirements.txt
aiohttp==3.6.2
astroid==2.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
et-xmlfile==1.0.1
flake8==3.8.2
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
openpyxl==3.0.3
peewee==3.13.3
pycodestyle==2.6.0
pyflakes==2.2.0
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wrapt==1.12.1
yarl==1.4.2
zipp==1.2.0